refactor: separate create and generate to module (#59)
