+++ src/generate/helpers/checkProperty.ts
@@ -1,4 +1,4 @@
-import { packageType } from "../types/interface.js";
+import { packageType } from "../../types/interface.js";
 
 /**
  * Checks whether the given property exists or not
