+++ src/generate/helpers/checkReact.ts
@@ -1,4 +1,4 @@
-import { packageType } from "../types/interface.js";
+import { packageType } from "../../types/interface.js";
 
 /**
  * Check wether the project is a React project or not
