+++ src/generate/helpers/getRootPath.ts
