+++ src/generate/template/component.ts
