+++ src/tests/getRootPath.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import getRootPath from "./getRootPath";
+import getRootPath from "../generate/helpers/getRootPath.js";
 
 describe("Get Root Path", () => {
     it("should return empty string if it doesn't contain 'src'", () => {
