+++ tsconfig.json
@@ -4,7 +4,6 @@
         "strict": true,
         "moduleResolution": "nodenext",
         "target": "ES2022",
-        "sourceMap": true,
         "noUnusedLocals": true,
         "noUnusedParameters": true,
         "noImplicitReturns": true,
@@ -13,6 +12,6 @@
         "strictNullChecks": true,
         "outDir": "dist"
     },
-    "include": ["src/**/*", "mock.test.ts"],
-    "exclude": ["node_modules", "src/**/*.test.ts"]
+    "include": ["src/**/*"],
+    "exclude": ["node_modules", "src/tests"]
 }
