+++ vitest.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
     test: {
         exclude: ["**/node_modules/**", "**/dist/**"],
-        include: ["src/**/*.test.ts"],
+        include: ["src/tests/*test.ts"],
         coverage: {
             reporter: ["text", "json", "html"],
             reportsDirectory: "./coverage",
